refactor(register): No more inner class from the ClientRegister, split into TimelessAnimations
